Split up the layouts for JustBanksy, JustPlaces, and JustSCOTUS
