+++ app/src/layout/JustBanksy/PreparedStatements.css
@@ -1,0 +1,15 @@
+.banksy-statement-generator .statement-generator-phrase {
+    font-family: monospace;
+    color: gray;
+    box-sizing: border-box;
+    padding: 0.5em;
+    font-size: 1.25em;
+}
+
+.dark .banksy-statement-generator .statement-generator-phrase.active {
+    color: yellow;
+}
+
+.light .banksy-statement-generator .statement-generator-phrase.active {
+    color: orangered;
+}
