+++ app/src/layout/JustPlaces/JustPlaces.tsx
@@ -1,33 +1,23 @@
 import { IonIcon } from '@ionic/react';
 import { locationOutline } from 'ionicons/icons';
 import { useState } from 'react';
-import { HStack, Spacer, VStack } from 'reaction';
+import { HStack, VStack } from 'reaction';
 import justPlacesPlaceholder from 'src/assets/images/backgrounds/places.png';
 import logo from 'src/assets/images/icons/Just Places.png';
-import yellowPagesIcon from 'src/assets/images/icons/yellow pages.png';
 import LoadingAnimation from 'src/components/LoadingAnimation';
-import RatingComponent from 'src/components/RatingComponent';
-import SearchResult from 'src/components/SearchResult';
 import SearchBar from 'src/components/ui/SearchBar';
 import Sidebar from 'src/components/ui/Sidebar';
 import { useTitle } from 'src/HTMLHead';
 import { PlacesKit } from 'src/utils/JustSDK';
 import './JustPlaces.css';
-
-interface BusinessAgeBadgeProps {
-    age: number;
-}
-
-interface JustPlacesRatingProps {
-    rating: number;
-}
+import { PlaceSearchResult } from './PlaceSearchResult';
 
 /**
  * The page which displays the Just Places application.
  *
  * @returns The Just Places page.
  */
-function JustPlaces() {
+export default function JustPlaces() {
     // Search query and location from the titlebar's search bar
     const [searchQuery, setSearchQuery] = useState('');
     const [searchLocation, setSearchLocation] = useState('');
@@ -87,84 +77,3 @@ function JustPlaces() {
         </HStack>
     );
 }
-
-/**
- * A component which extends the base `SearchResult` component to display a search result
- * for a Yellow Pages SERP result.
- *
- * Note: Rather than applying each property using the `attr={value}` notation, the
- * spread operation should be used insted:
- *
- * ```
- * <PlaceSearchResult {...result} />
- * ```
- *
- * @param props - Properties belonging to `YellowPagesSearchResult` which contains
- * all the data from a Yellow Pages search.
- * @returns The search result component.
- */
-function PlaceSearchResult(props: YellowPagesSearchResult) {
-    return (
-        <SearchResult>
-            <HStack>
-                <h3 className='text-gradient-primary'>{props.business}</h3>
-                <Spacer />
-                <span>{props.openStatus}</span>
-            </HStack>
-            <HStack className='categories' justify='start'>
-                {props.categories.map(category => (
-                    <span>{category}</span>
-                ))}
-            </HStack>
-            <HStack justify='between'>
-                {props.rating.yellowPages && (
-                    <YellowPagesRating rating={props.rating.yellowPages} />
-                )}
-            </HStack>
-            <HStack justify='between'>
-                <VStack align='start'>
-                    <span>{props.phone}</span>
-                    <a href={props.website} className='text-gradient-link'>
-                        Website
-                    </a>
-                </VStack>
-                <ResultAddress {...props.address} />
-                {props.age && <BusinessAgeBadge age={props.age} />}
-            </HStack>
-        </SearchResult>
-    );
-}
-
-function YellowPagesRating({ rating }: JustPlacesRatingProps) {
-    return (
-        <HStack width='auto'>
-            <img src={yellowPagesIcon} alt='Yellow Pages Rating' style={{ height: '2rem' }} />
-            <RatingComponent rating={rating} color='yellow' />
-        </HStack>
-    );
-}
-
-function ResultAddress({ street, city, state, zip }: YellowPagesAddress) {
-    return (
-        <HStack align='start'>
-            <span>{street}</span>
-            <span>
-                {city}, {state} {zip}
-            </span>
-        </HStack>
-    );
-}
-
-function BusinessAgeBadge({ age }: BusinessAgeBadgeProps) {
-    return (
-        <HStack width='auto'>
-            <VStack className='business-age'>{age}</VStack>
-            <VStack className='business-age-label' align='start'>
-                <span>Years in</span>
-                <span>Business</span>
-            </VStack>
-        </HStack>
-    );
-}
-
-export default JustPlaces;
